fix: In CONTRIBUTING, link directly to bug and feature request templates.
Currently for bugs and feature requests, the CONTRIBUTING guide links to the generic blank issue page. But we now have templates. This pull request changes the link to the "choose" page.

Since the bug report template already asks for all the details, we don't have to repeat that here.

Finally, renamed "Suggest a feature" to "Request a feature" for consistency. Also ask the user to check if it has already been requested.

Signed-off-by: jefft0 <jeff@thefirst.org>